requirements.txt
# Core Flask application
flask==2.3.3
requests==2.31.0
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.0.1
urllib3==2.0.7
werkzeug==2.3.7

# Security
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# Monitoring
prometheus-flask-exporter==0.22.4
prometheus-client==0.19.0
sentry-sdk[flask]==1.40.0

# API Documentation
flask-swagger-ui==4.11.1

# AI Service APIs (pinned versions)
openai==2.8.1
anthropic==0.8.1
google-generativeai==0.8.5  # Keep for backward compatibility (older Veo code)
google-genai==1.51.0  # NEW SDK for Imagen 4.0 and Veo 3.1
transformers==4.35.2
torch==2.1.2

# Google Cloud Storage (for potential media uploads)
google-cloud-aiplatform>=1.56.0
google-genai>=0.8.0
pillow>=10.0.0

# Async support for media generation
aiohttp==3.9.1
asyncio-mqtt==0.16.2

# Media processing for mock video/image generation
imageio==2.37.2
imageio-ffmpeg==0.6.0

